==> 01-packages.Rmd <==
# Packages {#packages}
The 2020 IR uses tools developed by Utah DWQ and stored as R-packages. Both packages (wqTools & irTools) are available for review, download, or installation via GitHub.  
Please see the [wqTools](https://github.com/utah-dwq/wqTools) and [irTools](https://github.com/utah-dwq/irTools) GitHub pages for additional information.

## Package installs
```{r, eval=F}
install.packages("devtools")
devtools::install_github("utah-dwq/wqTools")
devtools::install_github("utah-dwq/irTools")
```

## Libraries
```{r}
library(wqTools)
library(irTools)
library(magrittr)
```
