==> 02-data-download.Rmd <==
# Data download from EPA Water Quality Portal {#data-download}

The majority of assessment data used in the IR are available via the EPA [Water Quality Portal](https://www.waterqualitydata.us/]). Additional data from 
internal and external data submissions are also included in the assessment. These data were either manually appended to WQP data prior to assessment or 
assessed separately. The period of record for the 2020 IR includes water years 2011-2018. From the IR Assessment Methods:

> DWQ uses water years to define the period of record. DWQ uses the same definition as the [U.S. Geologic Survey](https://water.usgs.gov/nwc/explain_data.html) and defines the water year as the 12-month period between October 1 and September 30 of the following year. For the 2018/2020 IR, the period of record is October 1, 2010 to September 30, 2018 (water years 2011-2018).


## Download sites
```{r, eval=FALSE}
downloadWQP(outfile_path='01_raw_data', zip=FALSE, retrieve="sites")
```
## Download narrowresult, activity, and detection/quantitation
```{r, eval=FALSE}
downloadWQP(outfile_path='01_raw_data',start_date='10/01/2010', end_date='09/30/2018', zip=TRUE, unzip=TRUE, retrieve=c("narrowresult", "activity", "detquantlim"))
```

## Identify characters in result value column
```{r, eval=FALSE}
nr = read.csv("01_raw_data\\narrowresult-2019-08-22.csv", na.strings=c("NA", ""))
dim(nr)
nr_sub = subset(nr, !is.na(ResultMeasureValue))
dim(nr_sub)

nr_sub$RMV_num=wqTools::facToNum(nr_sub$ResultMeasureValue)
table(is.na(nr_sub$RMV_num))

nr_na=subset(nr_sub, is.na(RMV_num))

table(droplevels(nr_na$CharacteristicName))
table(droplevels(nr_na$OrganizationIdentifier))
```

*There were just over 5000 records that could not be interpreted to numeric values. Replaced commas, <, >, J, etc.*

## Replacing special characters
```{r, eval=FALSE}
nr=within(nr, {
	gsub(",", "", ResultMeasureValue)
	gsub("< ", "", ResultMeasureValue)
	gsub("<", "", ResultMeasureValue)
	gsub(">", "", ResultMeasureValue)
	gsub(" J", "", ResultMeasureValue)
})
nr_sub = subset(nr, !is.na(ResultMeasureValue))
dim(nr_sub)
nr_sub$RMV_num=wqTools::facToNum(nr_sub$ResultMeasureValue)
table(is.na(nr_sub$RMV_num))
nr_na=subset(nr_sub, is.na(RMV_num))
table(droplevels(nr_na$CharacteristicName))
```

## Write NR back to file
```{r, eval=FALSE}
write.csv(nr, file='01_raw_data/narrowresult-2019-08-22-spec-chars-filled.csv', row.names=F)
```


## Import data to workspace
```{r, eval = F}
irdata <- readWQPFiles(file_select=FALSE,
            narrowresult_file = "01_raw_data\\narrowresult-2019-08-22-spec-chars-filled.csv",
            sites_file = "01_raw_data\\sites-2019-08-22.csv",
            activity_file = "01_raw_data\\activity-2019-08-22.csv",
            detquantlim_file = "01_raw_data\\detquantlim-2019-08-22.csv",
            orph_check = TRUE)
objects(irdata)
save(irdata, file = "irdata_wqp_raw.Rdata")
```

## Subset to sites w/ data
```{r, eval=FALSE}
irdata$sites=subset(irdata$sites, MonitoringLocationIdentifier %in% irdata$merged_results$MonitoringLocationIdentifier)
write.csv(irdata$sites, "01_raw_data\\sites-2019-08-22.csv", row.names=F)
```
