==> 05-lake-assessments.Rmd <==
# Lake assessments {#lake-asmnts}
Manually side-stepping much of the IR data process for lake profile and trophic assessments.

## Profile data
### Read data from WQP
*Note. Querying long term data for visualization and review purposes, but only assessing data within IR POR. For this analysis, limiting data to UTAHDWQ_WQX organization because these samples include the profile measurements required for lake assessments.*

```{r, read-profiles, eval = FALSE}
profile_params=c('Dissolved oxygen (DO)','Temperature, water','pH', 'Depth, data-logger (ported)')
profile_data=readWQP(type='narrowresult', siteType='Lake, Reservoir, Impoundment', organization='UTAHDWQ_WQX', characteristicName=profile_params)

profile_data=within(profile_data, {
	DataLoggerLine[DataLoggerLine %in% c("`", "N/A")]="NA"
})

lake_sites=readWQP(type='sites', siteType='Lake, Reservoir, Impoundment', organization='UTAHDWQ_WQX')
lake_sites=subset(lake_sites, MonitoringLocationIdentifier %in% profile_data$MonitoringLocationIdentifier)
lake_sites=assignAUs(lake_sites)
all(profile_data$MonitoringLocationIdentifier %in% lake_sites$MonitoringLocationIdentifier)
```

### Check parameter units
```{r, eval = FALSE}
knitr::kable(table(profile_data$CharacteristicName, profile_data$ResultMeasure.MeasureUnitCode))
profile_data=subset(profile_data, !ResultMeasure.MeasureUnitCode %in% c('mg', 'mg/m2'))
```

### Merge spatial info to data
```{r, lake-merge-spatial, eval = FALSE}
profile_data=merge(profile_data, lake_sites, all.x=T)
```

### Reformat & add needed columns
```{r, eval = FALSE}
profile_data=within(profile_data, {
	R3172ParameterName=CharacteristicName
	R3172ParameterName[R3172ParameterName=='Depth, data-logger (ported)']='Profile depth'
	IR_Value=as.numeric(ResultMeasureValue)
	IR_MLID=MonitoringLocationIdentifier
	IR_MLNAME=MonitoringLocationName
	IR_Lat=LatitudeMeasure
	IR_Long=LongitudeMeasure
	IR_Unit=ResultMeasure.MeasureUnitCode
})	
profile_data=subset(profile_data, !is.na(DataLoggerLine))
```


## Site map
```{r, eval = FALSE}
lake_sites=within(lake_sites, {
	MonitoringLocationTypeName_wqp=MonitoringLocationTypeName
	MonitoringLocationTypeName=ifelse(MonitoringLocationIdentifier %in% profile_data$MonitoringLocationIdentifier, 'Profile', 'Other')
})
buildMap(sites=lake_sites)
```

### Assign uses to profile data, subset to ALUs
```{r, lake-assign-uses, eval = FALSE}
sites_uses=(assignUses(lake_sites, flatten=T))[,c('MonitoringLocationIdentifier','BeneficialUse','R317Descrp','Water_Type','bu_class')]
names(sites_uses)[names(sites_uses)=='bu_class']='BEN_CLASS'
profile_data=merge(profile_data, sites_uses, all.x=T)
```

### Remove faulty depth profiles
*Plotting revealed profile events in Jordanelle and Deer Creek reservoirs with depths exceeding 300 m. These are being removed here via AU & date.*
```{r,eval = FALSE}
profile_data=subset(profile_data, !(ASSESS_ID=='UT-L-16020203-003_00' & ActivityStartDate==as.Date('2004-07-20')))
profile_data=subset(profile_data, !(ASSESS_ID=='UT-L-16020203-001_00' & ActivityStartDate==as.Date('2004-07-21')))
```

### Assigning AUs & uses to out-of-state sites (for non-assessment dataset)
```{r, eval = FALSE}
man_uses_3A=c('UTAHDWQ_WQX-4906960','UTAHDWQ_WQX-4906980','UTAHDWQ_WQX-5939770','UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640','UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685')
man_uses_3B=c('UTAHDWQ_WQX-5952430')
profile_data_asmnt=subset(profile_data, BeneficialUse %in% c("3A","3B") & AU_Type == 'Reservoir/Lake' & as.Date(ActivityStartDate)>as.Date('2010-10-01') & as.Date(ActivityStartDate)<as.Date('2018-09-30'))
profile_data_all=within(profile_data, {
	BeneficialUse[MonitoringLocationIdentifier %in% man_uses_3A] = '3A'
	BeneficialUse[MonitoringLocationIdentifier %in% man_uses_3B] = '3B'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4906960', 'UTAHDWQ_WQX-4906980')] = 'UT-L-16010201-003_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4906960', 'UTAHDWQ_WQX-4906980')] = 'Bear Lake'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5939770')] = 'UT-L-14040107-001_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5939770')] = 'Meeks Cabin Reservoir'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640','UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685')] = 'UT-L-14040106-021_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640','UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685')] = 'Flaming Gorge Reservoir'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5952430')] = 'UT-L-14070006-001_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5952430')] = 'Lake Powell'
	AU_Type[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4906960','UTAHDWQ_WQX-4906980','UTAHDWQ_WQX-5939770','UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640',
		'UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685','UTAHDWQ_WQX-5952430')] = 'Reservoir/Lake'
})
profile_data_all=subset(profile_data_all, BeneficialUse %in% c("3A","3B") & AU_Type == 'Reservoir/Lake')
```


### Assign criteria
```{r, assign-lake-criteria, eval = FALSE}
prof_crit=read.csv(file='P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/lake_prof_criteria.csv')
profile_data_asmnt=merge(profile_data_asmnt, prof_crit, all.x=T)
profile_data_all=merge(profile_data_all, prof_crit, all.x=T)
```


test=subset(profile_data_asmnt, IR_MLID=="UTAHDWQ_WQX-4923810" & ActivityStartDate=="2018-09-04")

### Run assessments
*Note - not accounting for Kens Lake site-specific temperature criterion here. Need to check in secondary review.*
```{r, eval = FALSE}
table(profile_data_asmnt$R3172ParameterName, exclude=NULL)
prof_asmnts=assessLakeProfiles(profile_data_asmnt)
table(prof_asmnts$profile_asmnts_mlid_param$R3172ParameterName, exclude=NULL)
table(profile_data_all$R3172ParameterName, exclude=NULL)
prof_asmnts_all=assessLakeProfiles(profile_data_all)
table(prof_asmnts_all$profile_asmnts_mlid_param$R3172ParameterName, exclude=NULL)
```

## Trophic data
### Read data from WQP
```{r, read-trophic-data, eval = FALSE}
trophic_params=c('Phosphate-phosphorus','Chlorophyll a', 'Chlorophyll a, uncorrected for pheophytin', 'Chlorophyll a, corrected for pheophytin','Depth, Secchi disk depth')
trophic_data=readWQP(type='result', siteType='Lake, Reservoir, Impoundment', organization='UTAHDWQ_WQX', characteristicName=trophic_params)

trophic_data=merge(trophic_data, lake_sites, all.x=T)
trophic_data=within(trophic_data, {
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4906960', 'UTAHDWQ_WQX-4906980')] = 'UT-L-16010201-003_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4906960', 'UTAHDWQ_WQX-4906980')] = 'Bear Lake'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5939770')] = 'UT-L-14040107-001_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5939770')] = 'Meeks Cabin Reservoir'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640','UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685')] = 'UT-L-14040106-021_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640','UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685')] = 'Flaming Gorge Reservoir'
	ASSESS_ID[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5952430')] = 'UT-L-14070006-001_00'
	AU_NAME[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-5952430')] = 'Lake Powell'
	AU_Type[MonitoringLocationIdentifier %in% c('UTAHDWQ_WQX-4906960','UTAHDWQ_WQX-4906980','UTAHDWQ_WQX-5939770','UTAHDWQ_WQX-4938610','UTAHDWQ_WQX-4938620','UTAHDWQ_WQX-4938640',
		'UTAHDWQ_WQX-4938650','UTAHDWQ_WQX-4938680','UTAHDWQ_WQX-4938660','UTAHDWQ_WQX-4938685','UTAHDWQ_WQX-5952430')] = 'Reservoir/Lake'
})
trophic_data=trophic_data[!is.na(trophic_data$ASSESS_ID),]
```

### Check fractions, units, methods
```{r, eval = FALSE}
with(trophic_data, {
	knitr::kable(table(CharacteristicName, ResultSampleFractionText, exclude=NULL))
})
```
*Remove dissolved phosphorus*
```{r, eval = FALSE}
trophic_data=subset(trophic_data, !(CharacteristicName=='Phosphate-phosphorus' & ResultSampleFractionText=='Dissolved'))
trophic_data=trophic_data[!(trophic_data$CharacteristicName=='Phosphate-phosphorus' & trophic_data$ResultSampleFractionText=='Dissolved'),]
with(trophic_data, {
	knitr::kable(table(CharacteristicName, ResultSampleFractionText, exclude=NULL))
})
with(trophic_data, {
	knitr::kable(table(CharacteristicName, ResultMeasure.MeasureUnitCode, exclude=NULL))
})
with(trophic_data, {
	knitr::kable(table(ResultAnalyticalMethod.MethodName, CharacteristicName , exclude=NULL))
})
with(trophic_data[trophic_data$CharacteristicName %in% c('Chlorophyll a', 'Chlorophyll a, uncorrected for pheophytin', 'Chlorophyll a, corrected for pheophytin'),], {
	knitr::kable(table(ResultAnalyticalMethod.MethodName, CharacteristicName , exclude=NULL))
})
```

### Compare chl-a parameters
```{r, eval = FALSE}
boxplot((ResultMeasureValue+0.001)~CharacteristicName, trophic_data[trophic_data$CharacteristicName %in% c('Chlorophyll a', 'Chlorophyll a, uncorrected for pheophytin', 'Chlorophyll a, corrected for pheophytin'),], log='y', ylab='Chlorophyll a (ug/L)', xlab='')
```

### Translate all chl-a parameters to single CharacteristicName
```{r, eval = FALSE}
trophic_data=within(trophic_data, {
	CharacteristicName[CharacteristicName %in% c('Chlorophyll a', 'Chlorophyll a, uncorrected for pheophytin', 'Chlorophyll a, corrected for pheophytin')] = 'Chlorophyll a'
})
```

### Fill non-detects w/ 1/2 detection, set zero values to 0.001 for TSI/plotting
```{r, eval = FALSE}
trophic_data=within(trophic_data, {
	ResultMeasureValue[which(ResultDetectionConditionText=='Not Detected')] = DetectionQuantitationLimitMeasure.MeasureValue[which(ResultDetectionConditionText=='Not Detected')]/2
})
knitr::kable(table(trophic_data$CharacteristicName, trophic_data$ResultMeasureValue==0))
trophic_data$ResultMeasureValue[trophic_data$ResultMeasureValue==0]=0.001

trophic_data=calcTSI(trophic_data, value_var='ResultMeasureValue')
boxplot(TSI~CharacteristicName, trophic_data)
```

### Tier 2 assessments
#### Phytoplankton
```{r, eval = FALSE}
phyto_data=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/lake_phytoplankton.xlsx', 'data'))
phyto_sites=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/lake_phytoplankton.xlsx', 'sites'))
names(phyto_data)[names(phyto_data) %in% names(phyto_sites)]
phyto_data=merge(phyto_data, phyto_sites, all.x=T)
phyto_data=subset(phyto_data, !is.na(LatitudeMeasure) & !is.na(LongitudeMeasure)  & as.Date(ActivityStartDate)>as.Date('2010-10-01') & as.Date(ActivityStartDate)<as.Date('2018-09-30'))
phyto_data=assignAUs(phyto_data)
head(phyto_data)
```

##### Calculate cyanobacteria relative abundance
```{r, eval = FALSE}
phyto_data_division=reshape2::dcast(data=phyto_data, MonitoringLocationID+MonitoringLocationName+ActivityStartDate+AU_NAME+ASSESS_ID+ResultUnit~BiologicalIndividualID, value.var='ResultValue', fun.aggregate=sum)
phyto_data_division=within(phyto_data_division, {
	TotalAbd=Bacillariophyta+Chlorophyta+Chrysophyta+Cryptophyta+Cyanophyta+Dinophyta
	CyanoRA=Cyanophyta/TotalAbd
	CBdom=ifelse(CyanoRA>0.5, 1, 0)
})
```

##### Assess cyanobacteria RA
```{r, eval = FALSE}
samp_count=aggregate(CBdom~AU_NAME+ASSESS_ID, data=phyto_data_division, FUN='length')
names(samp_count)[names(samp_count)=='CBdom']='PhytoSampleCount'
cbdom_count=aggregate(CBdom~AU_NAME+ASSESS_ID, data=phyto_data_division, FUN='sum')
names(cbdom_count)[names(cbdom_count)=='CBdom']='CBdomCount'
cbdom_asmnt=merge(cbdom_count, samp_count)
cbdom_asmnt=within(cbdom_asmnt, {
	CBdomProp=CBdomCount/PhytoSampleCount
	CBdomAsmnt=NA
	CBdomAsmnt[CBdomProp>0.25 & PhytoSampleCount>=2] ='NS'
	CBdomAsmnt[CBdomProp<=0.25 & PhytoSampleCount>=2] ='FS'
})
```

#### Chl a TSI
```{r, eval = FALSE}
chla_tsi_data=subset(trophic_data, CharacteristicName =='Chlorophyll a' & as.Date(ActivityStartDate)>as.Date('2010-10-01') & as.Date(ActivityStartDate)<as.Date('2018-09-30'))
chla_tsi_data=within(chla_tsi_data, {
	tsi50=ifelse(TSI>50, 1, 0)
})
samp_count=aggregate(tsi50~AU_NAME+ASSESS_ID, chla_tsi_data, FUN='length')
tsi50_count=aggregate(tsi50~AU_NAME+ASSESS_ID, chla_tsi_data, FUN='sum')
names(samp_count)[names(samp_count)=='tsi50']='ChlaSampleCount'
names(tsi50_count)[names(tsi50_count)=='tsi50']='tsi50_count'
chlaTSIasmnt=merge(tsi50_count, samp_count)
chlaTSIasmnt=within(chlaTSIasmnt,{
	chlaTSIasmnt=NA
	tsi50_prop=tsi50_count/ChlaSampleCount
	chlaTSIasmnt[tsi50_prop>0.5 & ChlaSampleCount>2]='NS'
	chlaTSIasmnt[tsi50_prop<=0.5 & ChlaSampleCount>2]='FS'
})
```

#### Merge tier II assessments
```{r, eval = FALSE}
tier2asmnt=merge(chlaTSIasmnt, cbdom_asmnt, all=T)
with(tier2asmnt, {table(chlaTSIasmnt, CBdomAsmnt)})
tier2asmnt=within(tier2asmnt, {
	au_param_Cat2020=NA
	au_param_Cat2020[chlaTSIasmnt=='NS' & CBdomAsmnt=='NS']='NS'
	au_param_Cat2020[chlaTSIasmnt=='FS' & CBdomAsmnt=='FS']='FS'
	au_param_Cat2020[is.na(au_param_Cat2020)]='ID'
	R3172ParameterName='Lakes tier II'
})
tier2asmnt=tier2asmnt[,c('AU_NAME','ASSESS_ID','R3172ParameterName','au_param_Cat2020')]
tier2asmnt=subset(tier2asmnt, au_param_Cat2020!='ID')
```

### ID new listings and potential delistings
```{r, eval = FALSE}
prev_asmnt=read.csv('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/previous_lake_assessments.csv')
prev_asmnt_au_param=rollUp(data=list(prev_asmnt), group_vars=c('ASSESS_ID','BeneficialUse','R3172ParameterName'), cat_var='PreviousAssessment', print=F, expand_uses=F)
names(prev_asmnt_au_param)[names(prev_asmnt_au_param)=='AssessCat']='PrevCat'
asmnt2020_au_param=rollUp(data=list(prof_asmnts$profile_asmnts_mlid_param), group_vars=c('ASSESS_ID','AU_NAME','BeneficialUse','R3172ParameterName'), cat_var='IR_Cat', print=F, expand_uses=F)
names(asmnt2020_au_param)[names(asmnt2020_au_param)=='AssessCat']='au_param_Cat2020'
uses=unique(asmnt2020_au_param[,c('ASSESS_ID', 'BeneficialUse')])
tier2asmnt=merge(tier2asmnt, uses, all.x=T)
asmnt2020_au_param=rbind(asmnt2020_au_param, tier2asmnt)

comb_asmnt_au=merge(prev_asmnt_au_param, asmnt2020_au_param, all.y=T, by=c('ASSESS_ID','BeneficialUse', 'R3172ParameterName'))
knitr::kable(table(comb_asmnt_au$PrevCat, comb_asmnt_au$au_param_Cat2020, exclude=NULL))

comb_asmnt_au=within(comb_asmnt_au, {
	new_listing=NA
	new_listing[PrevCat %in% c('ID','FS') & au_param_Cat2020=='NS'] = 'Y'
	new_listing[is.na(PrevCat) & au_param_Cat2020=='NS'] = 'Y'
	pot_delist=NA
	pot_delist[PrevCat=='NS' & au_param_Cat2020=='FS'] = 'Y'
	PrevCat[is.na(PrevCat)]='FS'
})

knitr::kable(table(comb_asmnt_au$new_listing))
new_listings=subset(comb_asmnt_au, new_listing=='Y')
knitr::kable(new_listings, row.names=F)
pot_delist=subset(comb_asmnt_au, pot_delist=='Y')
knitr::kable(pot_delist, row.names=F)
```

### Write out assessments
```{r, eval = FALSE}
writexl::write_xlsx(list(lake_profile_asmnts=comb_asmnt_au),
		path = 'P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/lake_profile_asmnts.xlsx', format_headers=F, col_names=T)
```

## Save objects
```{r, eval = FALSE}
#save.image('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/lake_assessments.Rdata')
save(comb_asmnt_au, trophic_data, prof_asmnts_all, phyto_data, file='P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/05_lake_assessments/lake_data_for_app.Rdata')
```
