==> 08-export-assessments.Rmd <==
# Gather & export preliminary assessments & data {#export-asmnts}

## Gather & bind assessments
### Internal assessments
```{r}
lake_profs_assessed=as.data.frame(readxl::read_excel('lake_profile_asmnts.xlsx', 'lake_profile_asmnts'))
# lake_profs_assessed=comb_asmnt_au
names(lake_profs_assessed)[names(lake_profs_assessed)=='au_param_Cat2020']='IR_Cat'
table(lake_profs_assessed$R3172ParameterName)
lake_profs_assessed=within(lake_profs_assessed, {
	R3172ParameterName=as.character(R3172ParameterName)
	R3172ParameterName[R3172ParameterName=='Dissolved oxygen (DO)']='Minimum Dissolved Oxygen'
	R3172ParameterName[R3172ParameterName=='Temperature, water']='Max. Temperature'
})
prelim_asmnts=plyr::rbind.fill(conventionals_assessed,lake_tds_assessed,ss_mean_tds_assessed,toxics_assessed, assessed_ecoli$ecoli_mlid_asmnts,lake_profs_assessed[,!names(lake_profs_assessed) %in% c('PrevCat','au_param_Cat2020','pot_delist','new_listing')])

# prelim_asmnts=plyr::rbind.fill(toxics_and_conventionals_assessments$conventionals_assessed, toxics_and_conventionals_assessments$toxics_assessed,lake_assessments$lake_tds_assessed, site_specific_tds_assessments$ss_mean_tds_assessed,assessed_ecoli$ecoli_mlid_asmnts,lake_profs_assessed[,!names(lake_profs_assessed) %in% c('PrevCat','au_param_Cat2020','pot_delist','new_listing')])
```

### One parameter-site rejection from secondary review
One site along the Provo River was rejected for dissolved oxygen readings because the sample pulls from the bottom of Deer Creek Reservoir (below dam).
```{r}
reject = data.frame(IR_MLID="UTAHDWQ_WQX-5913210",R3172ParameterName="Minimum Dissolved Oxygen",Remove = "1")
prelim_asmnts = merge(prelim_asmnts, reject, all.x = TRUE)
prelim_asmnts = subset(prelim_asmnts, is.na(prelim_asmnts$Remove))
prelim_asmnts = prelim_asmnts[,!names(prelim_asmnts)%in%c("Remove")]
```

### External assessments
This section adds the Great Salt Lake egg selenium, fish tissue mercury, high frequency dissolved oxygen, Snowbird, and E.coli advisory and sampling assessments. Note that per our meeting with Nick von Stackelberg on dissolved oxygen holding capacity in high elevation waters, we are only assessing the 30-day average this cycle for the high frequency assessment.

Additionally, the Snowbird assessments occur at sites that are not included within the master site file because they did not come through the WQP.

External assessments occurring at the site-scale have their AU's re-assigned to take into account any resegmentation differences following secondary review. AU-scale external assessments affected by resegmentation need these edits made in the workbook rather than in the code.
```{r, eval = F}
# Great Salt Lake Egg Selenium
gsl_egg_se_assessed=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/gsl_egg_se/gsl_egg_se_assessment.xlsx', 'asmnt'))
names(gsl_egg_se_assessed)[names(gsl_egg_se_assessed)=='AssessCat']='IR_Cat'

# Fish Tissue Mercury
fish_hg_assessed=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/fish_tissue/3_draft_fishtissue_hgassessment_v2_ef.xlsx', 'asmnt'))
fish_hg_assessed=fish_hg_assessed[,!names(fish_hg_assessed) %in% 'Category to Report']
names(fish_hg_assessed)[names(fish_hg_assessed)=='AssessCat']='IR_Cat'

# High Frequency DO
hfdo_assessed=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/hfdo/hfdo_assessed_review.xlsx', 'HFDO_asmnts'))
hfdo_assessed = subset(hfdo_assessed, hfdo_assessed$AsmntAggPeriod==30)
### BPJ Per Erica's review: remove Silver Creek assessment because it is not in the main channel and the site is instead just below the WWTP outfall.
hfdo_assessed = hfdo_assessed[!hfdo_assessed$IR_MLID%in%c("UTAHDWQ_WQX-4926792"),]
hfdo_assessed$R3172ParameterName='High frequency DO'
hfdo_assessed=irTools::rollUp(list(hfdo_assessed), group_vars = c("BEN_CLASS","IR_MLID","IR_MLNAME","IR_Lat","IR_Long","ASSESS_ID", "AU_NAME", "R3172ParameterName","BeneficialUse"), expand_uses=F, print=F)
hfdo_assessed=within(hfdo_assessed, {
	IR_Lat=wqTools::facToNum(IR_Lat)
	IR_Long=wqTools::facToNum(IR_Long)
})
names(hfdo_assessed)[names(hfdo_assessed)=='AssessCat']='IR_Cat'

# Ecoli advisories
ecoli_adv_assessed=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/ecoli/3_draft_ecoliadvisory_assessment_v3_ef.xlsx', 'Sheet1'))
names(ecoli_adv_assessed)[names(ecoli_adv_assessed)=='AssessCat']='IR_Cat'

# SNOWBIRD
snowbird_assessed=as.data.frame(readxl::read_excel('P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/snowbird/snowbird_asmnt.xlsx', 'asmnt'))
names(snowbird_assessed)[names(snowbird_assessed)=='AssessCat']='IR_Cat'
snowbird_assessed=within(snowbird_assessed, {
	IR_Lat=wqTools::facToNum(IR_Lat)
	IR_Long=wqTools::facToNum(IR_Long)
})
snowbird_assessed = snowbird_assessed[,!names(snowbird_assessed)%in%c("Mgmt_Unit")]
snowbird_assessed$R3172ParameterName[snowbird_assessed$R3172ParameterName=="Temperature"] = "Max. Temperature"

# HABS
habs_assessed = as.data.frame(readxl::read_excel("P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/hab/3_draft_habassessment_v3_kf_eh.xlsx", "rolledUp_AU"))
habs_assessed = subset(habs_assessed, is.na(habs_assessed$`Remove?`))
au_bu_wmu = unique(acc_data_criteria[!is.na(acc_data_criteria$BEN_CLASS),c("ASSESS_ID","BEN_CLASS","Mgmt_Unit","AU_Type")])
habs_assessed = merge(habs_assessed, au_bu_wmu, all.x = TRUE)
habs_assessed$IR_MLID = paste0(habs_assessed$AU_NAME,"-HABs")
habs_assessed$IR_MLNAME = paste0(habs_assessed$AU_NAME,"-HABs")
habs_assessed$BeneficialUse = "2A"
habs_assessed$BeneficialUse = ifelse(grepl("2B", habs_assessed$BEN_CLASS),"2B","2A")
names(habs_assessed)[names(habs_assessed)=='AssessCat']='IR_Cat'
habs_assessed = habs_assessed[,!names(habs_assessed)%in%c("Remove?","AU_NAME_OLD")]
names(habs_assessed)[names(habs_assessed)=="ASSESS_ID_OLD"] = "ASSESS_ID2016"

# Bugs
bugs_assessed=as.data.frame(readxl::read_excel("P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/macroinvertebrates/1_draftanalysis_macroinvert_oeassessment_AssessNotes_v6_ef_eh.xlsx", '18-20 final assessments'))
bugs_assessed = bugs_assessed[,c("ASSESS_ID_OLD","ASSESS_ID","OriginalAssessCat","ResegmentNote","IR Note1","2016AssessCat","Category to Report")]
names(bugs_assessed) = c("ASSESS_ID2016","ASSESS_ID","OrigAssessCat_Bugs","EH_AUReseg_Note","EF_KF_BH_Note_Bugs","2016AssessCat_Bugs","IR_Cat")
bugs_assessed$R3172ParameterName = "Macroinvertebrates"
# Get bug beneficial uses
bugs_latlong=as.data.frame(readxl::read_excel("P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/99_external_assessments/macroinvertebrates/1_draftanalysis_macroinvert_oeassessment_AssessNotes_v5_ef_eh.xlsx", '18-20 data assessments'))
bugsites_latlong = unique(bugs_latlong[!is.na(bugs_latlong$DD_LAT_Y),c("DD_LAT_Y","DD_LONG_X")])
au_info = assignAUs(bugsites_latlong, lat = "DD_LAT_Y", long = "DD_LONG_X")
au_info = subset(au_info, !is.na(au_info$ASSESS_ID))
au_bu_info = assignUses(au_info,lat = "DD_LAT_Y", long = "DD_LONG_X")
names(au_bu_info)[names(au_bu_info)=="BeneficialUse"] = "BEN_CLASS"
au_bu_info = unique(au_bu_info[,!names(au_bu_info)%in%c("DD_LAT_Y","DD_LONG_X")])
bugs_assessed = merge(bugs_assessed, au_bu_info, all.x = TRUE)
bugs_assessed = subset(bugs_assessed, !bugs_assessed$ASSESS_ID%in%c("undefined"))
bugs_assessed$BeneficialUse = NA
bugs_assessed$BeneficialUse = gsub("1C,|2A,|2B,|4","",bugs_assessed$BEN_CLASS)
bugs_assessed$BeneficialUse = gsub(",$","",bugs_assessed$BeneficialUse)
bugs_assd_flat = tidyr::separate(data = bugs_assessed, col = BeneficialUse, into = c("BenUse1", "BenUse2"), sep = ",")
bugs_assd_flat = tidyr::pivot_longer(bugs_assd_flat, cols = c("BenUse1","BenUse2"), values_to = "BeneficialUse")
bugs_assd_flat = bugs_assd_flat[!is.na(bugs_assd_flat$BeneficialUse),!names(bugs_assd_flat)%in%c("name")]
bugs_assd_flat$IR_MLID = paste0(bugs_assd_flat$AU_NAME,"-Bugs")
bugs_assd_flat$IR_MLNAME = paste0(bugs_assd_flat$AU_NAME,"-Bugs")
rm(au_info, au_bu_info, bugs_latlong, bugsites_latlong)

external_asmnts=plyr::rbind.fill(gsl_egg_se_assessed, fish_hg_assessed, hfdo_assessed, ecoli_adv_assessed, snowbird_assessed, bugs_assd_flat, habs_assessed)

## Assign newly resegmented AU's to external assessments
ext_assignaus = subset(external_asmnts, !is.na(external_asmnts$IR_Lat))
ext_noassing = subset(external_asmnts, is.na(external_asmnts$IR_Lat))
ext_assignaus = unique(ext_assignaus[,!names(ext_assignaus)%in%c("ASSESS_ID2016","AU_Type","Category to Report","2016AssessCat_Bugs","AssessCat_OLD","Mgmt_Unit")])
names(ext_assignaus)[names(ext_assignaus)=="ASSESS_ID"] = "ASSESS_ID2016"
names(ext_assignaus)[names(ext_assignaus)=="AU_NAME"] = "AU_NAME2016"
ext_aus = assignAUs(data = ext_assignaus, lat = "IR_Lat", long ="IR_Long")

external_asmnts = plyr::rbind.fill(ext_noassing, ext_aus)

save(external_asmnts, file = "external_assessments.RData") # latest version: Oct 9, 2020, includes both rounds of AU resegmentations and HAB edits
```

#### Combined assessments
```{r}
load("external_assessments.RData")
prelim_asmnts=plyr::rbind.fill(prelim_asmnts, external_asmnts)
prelim_asmnts$pol_ind=ifelse(prelim_asmnts$ParameterGroupName=='POLLUTION INDICATORS' | prelim_asmnts$R3172ParameterName=='Lakes tier II', 'Y', 'N')
prelim_asmnts$pol_ind[is.na(prelim_asmnts$pol_ind)]='N'
any(is.na(prelim_asmnts$IR_Cat))
```

### Assign review metadata
#### WMUs and AU Types
```{r}
# au_wmu = readxl::read_xlsx("P:\\WQ\\Integrated Report\\IR_2020\\2020-IR-assessments\\spatial_files\\au_reseg\\site_au_reseg.xlsx", sheet = "AU WMU")## this is a local file. Change before public use.
load("au_wmu.Rdata")
prelim_asmnts = prelim_asmnts[,!names(prelim_asmnts)%in%c("Mgmt_Unit")]
prelim_asmnts = merge(prelim_asmnts, au_wmu, all.x = TRUE)
any(is.na(prelim_asmnts$Mgmt_Unit))
na_wmu = subset(prelim_asmnts, is.na(prelim_asmnts$Mgmt_Unit))

autypes = unique(data.frame(ASSESS_ID = au_poly$ASSESS_ID, AU_Type = au_poly$AU_Type))
prelim_asmnts = prelim_asmnts[,!names(prelim_asmnts)%in%c("AU_Type")]
prelim_asmnts = merge(prelim_asmnts,autypes, all.x = TRUE)

```

#### Lakes
```{r}
prelim_asmnts$lake=ifelse(prelim_asmnts$AU_Type=='Reservoir/Lake' | prelim_asmnts$BeneficialUse=='5A', 'Y', 'N')
table(prelim_asmnts$lake, exclude=NULL)
```
#### BPJ Changes to Site-Level Assessments
```{r}
sjr_thallium = subset(prelim_asmnts, prelim_asmnts$ASSESS_ID%in%c("UT14080205-001_00","UT14080201-009_00")&prelim_asmnts$R3172ParameterName=="Thallium")
sjr_thallium$IR_Cat_BPJ = "IDEX"

millrace_copper_zinc = subset(prelim_asmnts, prelim_asmnts$ASSESS_ID%in%c("UT16020201-012_00")&prelim_asmnts$R3172ParameterName%in%c("Copper", "Zinc")&prelim_asmnts$IR_Cat=="NS")
millrace_copper_zinc$IR_Cat_BPJ = "IDEX"

provoriver_do = subset(prelim_asmnts, prelim_asmnts$ASSESS_ID%in%c("UT16020203-001_00")&prelim_asmnts$R3172ParameterName%in%c("Minimum Dissolved Oxygen")&prelim_asmnts$IR_Cat=="NS")
provoriver_do$IR_Cat_BPJ = "IDEX"

cat3s = plyr::rbind.fill(sjr_thallium, millrace_copper_zinc, provoriver_do)

prelim_asmnts = merge(prelim_asmnts, cat3s, all.x = TRUE)

names(prelim_asmnts)[names(prelim_asmnts)=="IR_Cat"] = "IR_Cat_ORIG"
prelim_asmnts$IR_Cat = ifelse(is.na(prelim_asmnts$IR_Cat_BPJ), prelim_asmnts$IR_Cat_ORIG, prelim_asmnts$IR_Cat_BPJ)
```

#### Reviewer export file
```{r}
compiled_data = irTools::composeExport(prepped_data = prepped_data, toxics_assessed = toxics_assessed, conventionals_assessed = conventionals_assessed, create_workbooks = FALSE)

# If starting from prelim_asmnts.RData
# compiled_data = irTools::composeExport(prepped_data = prepped_data, toxics_assessed = toxics_and_conventionals_assessments$toxics_assessed, conventionals_assessed = toxics_and_conventionals_assessments$conventionals_assessed, create_workbooks = TRUE, params = "Minimum Dissolved Oxygen", include_rejected = TRUE)

save(compiled_data, file = "reviewer_export_data.Rdata")
```

#### Export all site-use-parameter assessments
```{r}
# Rename some objects to make them easier to remember and use
assessed_dat_crit_flat = prepped_data$acc_data
assessed_dat_allcols = subset(acc_data, acc_data$ResultIdentifier%in%unique(assessed_dat_crit_flat$ResultIdentifier))

save(acc_data_criteria, assessed_dat_crit_flat,assessed_dat_allcols,group_vars, lake_tds_assessed, lake_tds_exc, lake_profs_assessed, ss_mean_tds_assessed, ss_mean_tds_exc, conventionals_assessed, conventionals_exc, toxics_assessed, toxics_exc, prepped_data, assessed_ecoli, prelim_asmnts, compiled_data, file='prelim_asmnts_data.Rdata')
```

## Secondary review input files
#### Permits
```{r, eval = FALSE, echo = FALSE}
# permits=as.data.frame(readxl::read_excel("P:/WQ/Integrated Report/IR_2020/2020-IR-assessments/assessment/08_secondary_reviews/Permits/Current UPDES permits w_WLAs.xlsx", 'permits'))
# save(permits, file = "current_updes_permits.RData")

```

```{r, permits}
load("current_updes_permits.RData")
ut_fac1=wqTools::readECHO_fac(p_st="ut", p_act="y")
fac_coords1=do.call(rbind.data.frame,ut_fac1$geometry$coordinates)
names(fac_coords1)=c("dec_long","dec_lat")
fac_coords1=data.frame(ut_fac1$properties[,c("SourceID","CWPName","CWPFacilityTypeIndicator")], (fac_coords1))

permits2=subset(permits, !PermitID %in% ut_fac1$properties$SourceID)
ut_fac2=wqTools::readECHO_fac(p_pid=permits2$PermitID)
fac_coords2=do.call(rbind.data.frame,ut_fac2$geometry$coordinates)
names(fac_coords2)=c("dec_long","dec_lat")
fac_coords2=data.frame(ut_fac2$properties[,c("SourceID","CWPName","CWPFacilityTypeIndicator")], (fac_coords2))

ut_fac=rbind(fac_coords1, fac_coords2)
names(ut_fac)[names(ut_fac)=="SourceID"]="locationID"
names(ut_fac)[names(ut_fac)=="CWPName"]="locationName"
names(ut_fac)[names(ut_fac)=="CWPFacilityTypeIndicator"]="locationType"
names(ut_fac)[names(ut_fac)=="dec_long"]="LongitudeMeasure"
names(ut_fac)[names(ut_fac)=="dec_lat"]="LatitudeMeasure"

ut_fac=wqTools::assignAUs(ut_fac, lat = "LatitudeMeasure", long = "LongitudeMeasure")
ut_fac$priority=ifelse(ut_fac$locationID %in% permits$PermitID, 'Y', 'N')
permits_aus=subset(ut_fac, locationID %in% permits$PermitID)

prelim_asmnts$permit=ifelse(prelim_asmnts$ASSESS_ID %in% permits_aus$ASSESS_ID, 'Y','N')
table(prelim_asmnts$permit, exclude=NULL)

write.csv(file='ut_facilities.csv', ut_fac, row.names=F)
```

#### New listings (AU-parameter scale)
```{r}
prelim_asmnts$new_listing=NA
current_listings=read.csv(file='current_listings.csv')
prelim_asmnts=merge(prelim_asmnts, current_listings, all.x=T)
prelim_asmnts$new_listing=ifelse(prelim_asmnts$pol_ind=='N' & prelim_asmnts$IR_Cat=='NS' & is.na(prelim_asmnts$AU_EPACat), 'Y', 'N')
table(prelim_asmnts$new_listing, exclude=NULL)
prelim_asmnts$confirmed_listing=ifelse(prelim_asmnts$pol_ind=='N' & prelim_asmnts$IR_Cat=='NS' & !is.na(prelim_asmnts$AU_EPACat), 'Y', 'N')
table(prelim_asmnts$confirmed_listing, exclude=NULL)
prelim_asmnts=prelim_asmnts[,!names(prelim_asmnts) %in% c('AU_EPACat','PARAM_NAME')]
```

### Export files 
```{r}
site_use_param_asmnt=irTools::rollUp(list(prelim_asmnts), group_vars = c("BEN_CLASS","IR_MLID","IR_MLNAME","IR_Lat","IR_Long","ASSESS_ID","AU_NAME", "R3172ParameterName", "BeneficialUse", "pol_ind", "Mgmt_Unit","lake","permit","new_listing"), expand_uses=F)

site_use_param_asmnt=within(site_use_param_asmnt, {
	site_param_review=NA
	site_param_review[new_listing=='Y' & permit=='Y' & pol_ind=='N'] = 5
	site_param_review[new_listing=='Y' & permit=='N' & pol_ind=='N'] = 4
	site_param_review[new_listing=='N' & permit=='Y'] = 3
	site_param_review[new_listing=='N' & permit=='N'  & pol_ind=='N' & AssessCat=='IDEX' & !ASSESS_ID %in% subset(site_use_param_asmnt, site_param_review %in% c('New listing, permit', 'New listing'))$ASSESS_ID] = 2
	site_param_review[is.na(site_param_review)] = 1
})
table(site_use_param_asmnt$site_param_review, exclude=NULL)
au_revs=unique(site_use_param_asmnt[,c('ASSESS_ID','site_param_review')])
au_revs=aggregate(site_param_review~ASSESS_ID, au_revs, 'max')
names(au_revs)[names(au_revs)=='site_param_review']='au_rev'
site_use_param_asmnt=merge(site_use_param_asmnt, au_revs, all.x=T)

site_use_param_asmnt=within(site_use_param_asmnt, {
	AU_review=NA
	AU_review[au_rev==5]='New listing, permit'
	AU_review[au_rev==4]='New listing'
	AU_review[au_rev==3]='Permit'
	AU_review[au_rev==2]='New insufficient data w/ exc'
	AU_review[au_rev==1]='Review needed, low priority'
})
table(site_use_param_asmnt$AU_review, exclude=NULL)
write.csv(file='site_use_param_asmnt.csv', site_use_param_asmnt, row.names=F)
```

#### Secondary review input files
```{r, eval = F}
au_splits=as.data.frame(readxl::read_excel('site-use-param-asmnt-blank.xlsx', 'au-splits'))
au_reviews=as.data.frame(readxl::read_excel('site-use-param-asmnt-blank.xlsx', 'au-reviews'))
site_reviews=as.data.frame(readxl::read_excel('site-use-param-asmnt-blank.xlsx', 'site-reviews'))
record_reviews=as.data.frame(readxl::read_excel('site-use-param-asmnt-blank.xlsx', 'record-reviews'))
sw_reviews=as.data.frame(readxl::read_excel('site-use-param-asmnt-blank.xlsx', 'sw-reviews'))
# plyr::d_ply(site_use_param_asmnt, 'Mgmt_Unit',
# 	function(sdf) writexl::write_xlsx(list(
# 										`site-use-param-asmnt`=sdf, 
# 										`au-splits`=au_splits,
# 										`au-reviews`=au_reviews,
# 										`site-reviews`=site_reviews,
# 										`sw-reviews`=sw_reviews,
# 										`record-reviews`=record_reviews
# 	), path=paste0('secondary-review-input-', sdf$Mgmt_Unit[1], '.xlsx'), format_headers=F, col_names=T))

writexl::write_xlsx(list(
										`site-use-param-asmnt`=site_use_param_asmnt, 
										`au-splits`=au_splits,
										`au-reviews`=au_reviews,
										`site-reviews`=site_reviews,
										`sw-reviews`=sw_reviews,
										`record-reviews`=record_reviews
	), path=paste0('secondary-review-input-all.xlsx'), format_headers=F, col_names=T)
```

### Write asmntDashboard_data
*Currently excluding GSL egg, fish Hg, HFDO, macroinvertebrate, & lake profile data. These can be reviewed in other capacities.*
#### Plot data
```{r}
assessed_data=unique(prepped_data$acc_data[,c("CAS", "ResultIdentifier","ActivityStartDate", "IR_MLID", "IR_MLNAME", "ASSESS_ID", "AU_NAME", 
									"AU_Type", "BEN_CLASS", "IRParameterName", "R3172ParameterName", "IR_Value", "IR_Unit", "CriterionUnits",
									"IR_DetCond", "IR_Fraction", "IR_ActivityType", "IR_Lat", "IR_Long", 
									"DataLoggerLine", "ActivityRelativeDepthName", "ActivityDepthHeightMeasure.MeasureValue", 
									"R317Descrp", "ActivityDepthHeightMeasure.MeasureUnitCode")
					])
criteria=unique(prepped_data$acc_data[,c("CAS", "ActivityStartDate", "IR_MLID", "IR_MLNAME", "BeneficialUse", 
   								"R3172ParameterName", "CriterionUnits", "TargetFraction", "CriterionLabel", 
   								"CriterionType", "NumericCriterion","TableDescription", "ParameterQualifier", 
   								"FrequencyCombined", "FrequencyNumber", "FrequencyUnit", "TargetActivityType")
				])
criteria=subset(criteria, !BeneficialUse %in% c('CF', 'SUP'))
```

##### Build criteria labels
```{r}
criteria=within(criteria, {
	bu=ifelse(TableDescription=='LIST OF HUMAN HEALTH CRITERIA (CONSUMPTION)', paste0('HH-',BeneficialUse), BeneficialUse)
	frac=tolower(TargetFraction)
	els=ifelse(ParameterQualifier %in% c('early life stages are present','Fish Early Life Stages are Present'), 'ELS', NA)
	acc_chron=ifelse(CriterionLabel =='Acute', 'acute', 'chronic')
	ts=ifelse(R3172ParameterName=='Minimum Dissolved Oxygen' | R3172ParameterName=='E. coli', paste0(FrequencyNumber, ' day ', tolower(FrequencyCombined)), NA)
	ts=gsub('NA day', 'daily', ts)
	ts=gsub('average', 'avg', ts)
	ts=gsub('minimum', 'min', ts)
	label=paste(bu, R3172ParameterName, ts, els, acc_chron, frac) 
	label=gsub('Minimum Dissolved Oxygen', 'DO', label)
	label=gsub('Max. Temperature', 'Temperature', label)
	label=gsub(' NA', '', label)
	label=gsub('dissolved', '(diss)', label)
	label=gsub('total', '(tot)', label)
	label=gsub('none', '', label)
	label=gsub('daily maximum', 'max', label)
	label=gsub('geometric mean', 'geomean', label)
	label=ifelse(R3172ParameterName=='pH' | R3172ParameterName=='Total Dissolved Solids', paste(label, CriterionType), label)
})
head(subset(criteria, R3172ParameterName=='Total Dissolved Solids'))

criteria=criteria[,!names(criteria) %in% c('ts','acc_chron','els','frac','bu')]
save(assessed_data, criteria, file='asmntDashboard_data.Rdata')
```

#### Write assessment Excel file
```{r, eval=TRUE}
writexl::write_xlsx(list(
						site_use_param_asmnt=site_use_param_asmnt,
						prelim_asmnts=prelim_asmnts,
						lake_prof_asmnts=lake_profs_assessed,
						conventionals=subset(prepped_data$acc_data, AssessmentType=='Conventional'),
						toxics=subset(prepped_data$acc_data, AssessmentType=='Toxic'),
						supplement=subset(prepped_data$acc_data, AssessmentType=='All'),
						ecoli=subset(prepped_data$acc_data, AssessmentType=='Ecoli')
					),
		path = 'prelim_asmnts.xlsx', format_headers=F, col_names=T)
```


```{r, eval=F}
data=prepped_data$toxics
data$NumericCriterion=wqTools::facToNum(data$NumericCriterion)
data=within(data, {
	bu=ifelse(TableDescription=='LIST OF HUMAN HEALTH CRITERIA (CONSUMPTION)', paste0('HH-',BeneficialUse), BeneficialUse)
	frac=tolower(TargetFraction)
	els=ifelse(ParameterQualifier %in% c('early life stages are present','Fish Early Life Stages are Present'), 'ELS', NA)
	acc_chron=ifelse(CriterionLabel =='Acute', 'acute', 'chronic')
	ts=ifelse(R3172ParameterName=='Minimum Dissolved Oxygen' | R3172ParameterName=='E. coli', paste0(FrequencyNumber, ' day ', tolower(FrequencyCombined)), NA)
	ts=gsub('NA day', 'daily', ts)
	ts=gsub('average', 'avg', ts)
	ts=gsub('minimum', 'min', ts)
	label=paste(bu, R3172ParameterName, ts, els, acc_chron, frac) 
	label=gsub('Minimum Dissolved Oxygen', 'DO', label)
	label=gsub('Max. Temperature', 'Temperature', label)
	label=gsub(' NA', '', label)
	label=gsub('dissolved', '(diss)', label)
	label=gsub('total', '(tot)', label)
	label=gsub('none', '', label)
	label=gsub('daily maximum', 'max', label)
	label=gsub('geometric mean', 'geomean', label)
	label=ifelse(R3172ParameterName=='pH' | R3172ParameterName=='Total Dissolved Solids', paste(label, CriterionType), label)
})

library(plotly)
plot_data=unique(data[data$R3172ParameterName!='Pentachlorophenol' & data$R3172ParameterName!='Total Ammonia as N',c('NumericCriterion','hardness','R3172ParameterName','CriterionLabel')])
plot_data=dplyr::sample_frac(plot_data, 0.1)
plot_data$label=paste(plot_data$R3172ParameterName, tolower(plot_data$CriterionLabel))
hardness_dep_crit_plot=plot_ly(data=plot_data) %>%
	add_lines(x=~hardness, y=~NumericCriterion, color=~label) %>%
		layout(
			title = "Hardness dependent criteria",
			xaxis = list(title = "Hardness (mg/L)"),
			yaxis = list(side = 'left', title = 'Criterion (ug/L)')#,
			#legend = list(x = 0.01, y = 0.99)
		) %>% 
		config(displaylogo = FALSE,
			modeBarButtonsToRemove = c(
				'sendDataToCloud',
				'hoverClosestCartesian',
				'hoverCompareCartesian',
				'lasso2d',
				'select2d'
			)
		)
#hardness_dep_crit_plot
#save(file='C:/Users/jvander/Documents/R/dwq-R-tools-presentation/figures/hardness_dep_crit.Rdata', hardness_dep_crit_plot)
#htmlwidgets::saveWidget(hardness_dep_crit_plot, file='C:/Users/jvander/Documents/R/dwq-R-tools-presentation/figures/hardness_dep_crit.html')
api_create(hardness_dep_crit_plot, filename="hardness-dep-crit")
```
